requirements.txt
click==8.1.3
flake8==6.0.0
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.2.0
mypy-extensions==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
python-dotenv=1.0.0
requests~=2.31.0